chore: cleaned up env. put all data classes into data_classes.py. Extracted static variables into datatypes
